add draggability for cards
