+++ index.html
@@ -7,7 +7,8 @@
     <link rel="stylesheet" href="./css/style.css">
     <link rel="icon" type="image/png" sizes="32x32" href="/favicon-32x32.png">
     <title>Cards</title>
-
+    <script src = "https://code.jquery.com/jquery-1.10.2.js"></script>
+    <script src="https://code.jquery.com/ui/1.10.4/jquery-ui.js"></script>
     <script src="js/index.js" type="module" defer></script>
     <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
 </head>
@@ -43,10 +44,9 @@
             </form>
             
         </div>
-        <div class="p-3 border-dark rounded my-5 gap-5 fs-1 bg-opacity-50 bg-light">
-            <div id="root" class="row">
-                <p id="no-item" class="text-center">No items have been added</p>
-            </div>
+        <div class="p-3 rounded my-5 bg-opacity-50 bg-light">
+            <p id="no-item" class="text-center">No items have been added</p>
+            <ul id="root" class="row list-unstyled"></ul>
         </div>
     </main>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p" crossorigin="anonymous"></script>
