+++ js/functions/draggable.js
@@ -1,0 +1,5 @@
+function draggAndDropp(selector) {
+    $(selector).sortable();
+}
+
+export default draggAndDropp;
