+++ js/functions/initialRender.js
@@ -5,7 +5,6 @@ import { handleProps } from "./handleCardProps.js";
 
 async function initialRender() {
     const cards = await getCards();
-console.log(cards);
     if (cards.length !== 0) {
         document.getElementById('no-item').classList.add('d-none');
